Rename 'sequence' for 'raw_sequence' and support CDS region
The term 'raw_sequence' is exclusively used for sequence which may
contain linefeeds. This way, it's less confusing when we introduce the
concept of sequence length, which is the raw length minus the number of
line feeds.

Extract CDS assuming there's a field in the FASTA comment in the
'cds=a..b' format.