Compute site vacancy and silencing
The site vacancy is pretty solid for the sites that we fill in common
with the expected output, we get the same number of assigned miRNAs.
There's a couple of false positive that we need to investigate though.

The silencing is now computed by extracting CDS from the target fasta
file if it contains a 'cds=a..b' token in the sequence comment. Ideally
we would provide an external file with CDS to keep FASTA untouched.