sequence: Cache subsequence index at each position
For a relatively small memory cost, cache all the computed 'sequence_index'
for each position of the 'MirbookingSequence' instead of only the latest
position.

This reduces up drastically the cost of computing index for
'MirbookingTarget' as we repetitively ask for the same target site
during the execution of the algorithm.