Compute the silencing according the the Poisson-Binomial model
This approach combine two elements:

  - probability of observing k miRNA on a given target computed by the
    Poisson-Binomial distribution and estimated from occupant
    frequencies
  - regression done on the pSILAC data to have an idea of the mean
    silencing per number of affected sites

The silencing is basically a dot product between these two vectors
scaled up to the target quantity.

For the experiment output, it is distributed proportionally across the
occupancy of the miRNA such that the target silencing can be recovered
by doing a sum.

Reintroduce a column in 'mirbooking' output and in the dataframe
generated by the introspectable module and fix some bugs in the emitted
columns.