Rename 'PrecomputedScoreTable' into 'DefaultScoreTable' (fix #27)
With the addition of the enzymatic model, the responsibilities of the
score table goes beyond providing precomputed seed scores. Thus, the
renaming will reflect better the intent of providing the best possible
defaults for characterizing the miR::MRE interaction.

Add support for sparse score table (fix #9). It was first intended to be
a separate commit, but I cannot untangle the work anymore.

Update the table generator (fix #6) to generate sparse score tables.