broker: Don't keep track of position-wise concentrations
It is preferable to keep these values computed on-the-fly to have a
relatively small number of ODE to solve. Instead, we only keep track of
the total substrate concentrations.

With 'E' and 'S' defined this way, we don't need 'E0' and 'S0' anymore
as we can keep track of whole species concentration.