broker: Use a tighter lower bound on the site vacancy
We know for example that within a footprint window around a position, no
two occupants can overlap. Thus, the available concentration can be
lower-bounded by the sum of occupant concentrations.

However, this is incomplete because occupants further apart the
footprint window (i.e. a 3' + 5' positions downstream) can interfere
with the position, but not with the occupant the same distance upstream.