default-score-table: Add a term for enzyme turnover to kcat
Separate the kcat rate into kcleave modulated on the binding probability
of the cleavage site and krelease for the enzyme turnover.

Add the R^2 per constant in the fitting script so that we can appreciate
in more details the improvement. We now explain 30% of the catalytic
rates instead of 19% and 27% of Km instead of 20%.