Add support for a interaction blacklist filter (#82)
The filter relies on hashing MirbookingOccupant structures in order to
determine if an interaction is to be filtered out by the broker.

Move MirbookingOccupant-related routines into mirbooking-occupant.c and
add corresponding declaration in the header file to reuse them in the
binary.

Add an integration test to make sure that the blacklist filtering is
working.