Add tool for preparing input quantifications
Add basic support for GENCODE and miRBase FASTA format.

Add support for reading RSEM, HTSeq and TSV counts.

Use a geometric mean when replicated gene expression are provided with a
log(X+1) transformation.

The quantification is based on the assumption that the total
concentration of protein-coding transcripts and microRNA in the
cytoplasm is preserved across tissue types. We use a model of a HeLa S3
cell to estimate the number of species and the cytoplasmic volume.